+++ loader/loaders.py
@@ -135,7 +135,7 @@ def parse_proasis(input_string):
 def add_contacts(input_data,target,prot,mol):
     int_type = IntTypes()
     for interaction in input_data:
-        # Ignore Water mediasted interactions with Protein for now
+        # Ignore Water mediasted hypothesis with Protein for now
         if interaction['hetmoltype'] == 'WATER':
             continue
         res_name, res_num, chain_id = parse_proasis(interaction['dstrname'])
